requirements.txt
###### Requirements without Version Specifiers ######
djangorestframework


###### Requirements with Version Specifiers ######